+++ tests/data/WritingToReadonlyAttribute/ClassProperty.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Icanhazstring\PhpstanReadonlyPropertyExtension\Test\data\WritingToReadonlyAttribute;
 
-use Icanhazstring\PhpstanReadonlyPropertyExtension\Attribute\Readonly;
+use Icanhazstring\PhpstanReadonlyPropertyExtension\Attribute\IsReadonly;
 
 final class ClassProperty
 {
-    #[Readonly]
+    #[IsReadonly]
     public string $name;
 
     public function __construct(string $name)